FIX deepspeed dataloader for potential performances improvement
