+++ tutorials/distributed-ml/torch-scaling-test/deepspeed-config.yaml
@@ -1,13 +1,14 @@
 # Data and logging
-data_dir: tmp_data/
+data_dir: tmp_data/ #/p/scratch/intertwin/datasets/imagenet/ILSVRC2012/train/ # tmp_data/
 log_int: 10
 verbose: True
 nworker: 4 # num workers dataloader
 
 # Model
-batch_size: 64
+batch_size: 64 # micro batch size
 epochs: 3
 lr: 0.001
+shuff: False
 
 # Reproducibility
 rnd_seed: 10
