+++ tutorials/distributed-ml/torch-scaling-test/deepspeed_trainer.py
@@ -1,5 +1,5 @@
 """
-Scaling test of torch Distributed Data Parallel on Imagenet using Resnet.
+Scaling test of Microsoft Deepspeed on Imagenet using Resnet.
 """
 from typing import Optional
 import argparse
@@ -7,17 +7,19 @@ import sys
 import os
 from timeit import default_timer as timer
 import time
+import deepspeed
 
 import torch
 import torch.distributed as dist
-import torch.nn as nn
 import torch.nn.functional as F
+from torch.utils.data import DataLoader
+from torch.utils.data.distributed import DistributedSampler
 import torchvision
 
 from itwinai.parser import ArgumentParser as ItAIArgumentParser
 from itwinai.loggers import EpochTimeTracker
 
-from utils import seed_worker, imagenet_dataset, set_seed
+from utils import seed_worker, set_seed, imagenet_dataset
 
 
 def parse_params():
@@ -35,18 +37,14 @@ def parse_params():
     parser.add_argument('--nworker', type=int, default=0,
                         help=('number of workers in DataLoader '
                               '(default: 0 - only main)'))
-    parser.add_argument('--prefetch', type=int, default=2,
-                        help='prefetch data in DataLoader (default: 2)')
 
     # Model
-    parser.add_argument('--batch-size', type=int, default=64,
+    parser.add_argument('--batch-size', type=int, default=64, metavar='N',
                         help='input batch size for training (default: 64)')
-    parser.add_argument('--epochs', type=int, default=10,
+    parser.add_argument('--epochs', type=int, default=10, metavar='N',
                         help='number of epochs to train (default: 10)')
-    parser.add_argument('--lr', type=float, default=0.01,
+    parser.add_argument('--lr', type=float, default=0.01, metavar='LR',
                         help='learning rate (default: 0.01)')
-    parser.add_argument('--momentum', type=float, default=0.5,
-                        help='momentum in SGD optimizer (default: 0.5)')
     parser.add_argument('--shuff', action='store_true', default=False,
                         help='shuffle dataset (default: False)')
 
@@ -55,21 +53,25 @@ def parse_params():
                         help='seed integer for reproducibility (default: 0)')
 
     # Distributed ML
-    parser.add_argument('--backend', type=str, default='nccl',
-                        help='backend for parrallelisation (default: nccl)')
+    parser.add_argument('--backend', type=str, default='nccl', metavar='N',
+                        help='backend for parallelization (default: nccl)')
     parser.add_argument('--no-cuda', action='store_true', default=False,
                         help='disables GPGPUs')
+    parser.add_argument('--local_rank', type=int, default=-1,
+                        help='local rank passed from distributed launcher')
 
+    # parse to deepspeed
+    parser = deepspeed.add_config_arguments(parser)
     args = parser.parse_args()
-
     if args.verbose:
         args_list = [f"{key}: {val}" for key, val in args.items()]
         print("PARSED ARGS:\n", '\n'.join(args_list))
+
     return args
 
 
-def train(model, device, train_loader, optimizer, epoch, grank, gwsize, args):
-    model.train()
+def train(args, model, train_loader, optimizer, epoch, grank, gwsize):
+    device = model.local_rank
     t_list = []
     loss_acc = 0
     if grank == 0:
@@ -88,8 +90,8 @@ def train(model, device, train_loader, optimizer, epoch, grank, gwsize, args):
             print(
                 f'Train epoch: {epoch} [{batch_idx * len(data)}/'
                 f'{len(train_loader.dataset)/gwsize} '
-                f'({100.0 * batch_idx / len(train_loader):.0f}%)]\t\tLoss: '
-                f'{loss.item():.6f}')
+                f'({100.0 * batch_idx *len(data) / len(train_loader):.0f}%)]'
+                f'\t\tLoss: {loss.item():.6f}')
         t_list.append(timer() - t)
         loss_acc += loss.item()
     if grank == 0:
@@ -107,21 +109,23 @@ def main():
     if use_cuda and torch.cuda.device_count() > 0:
         is_distributed = True
 
+    # Limit # of CPU threads to be used per worker
+    # torch.set_num_threads(1)
+
     # Start the timer for profiling
     st = timer()
 
+    # Initializes the distributed backend
     if is_distributed:
-        # Initializes the distributed backend which will
-        # take care of synchronizing the workers (nodes/GPUs)
-        dist.init_process_group(backend=args.backend)
+        deepspeed.init_distributed(dist_backend=args.backend)
 
     # Set random seed for reproducibility
     torch_prng = set_seed(args.rnd_seed, use_cuda)
 
     if is_distributed:
-        # get job rank info - rank==0 master gpu
-        lwsize = torch.cuda.device_count()  # local world size - per run
+        # Get job rank info - rank==0 master gpu
         gwsize = dist.get_world_size()     # global world size - per run
+        lwsize = torch.cuda.device_count()  # local world size - per node
         grank = dist.get_rank()            # global rank - assign per run
         lrank = dist.get_rank() % lwsize     # local rank - assign per node
     else:
@@ -131,6 +135,7 @@ def main():
         grank = 0
         lrank = 0
 
+    # some debug
     if grank == 0:
         print('TIMER: initialise:', timer()-st, 's')
         print('DEBUG: local ranks:', lwsize, '/ global ranks:', gwsize)
@@ -138,33 +143,31 @@ def main():
         print('DEBUG: args.data_dir:', args.data_dir)
         print('DEBUG: args.log_int:', args.log_int)
         print('DEBUG: args.nworker:', args.nworker)
-        print('DEBUG: args.prefetch:', args.prefetch)
         print('DEBUG: args.batch_size:', args.batch_size)
         print('DEBUG: args.epochs:', args.epochs)
         print('DEBUG: args.lr:', args.lr)
-        print('DEBUG: args.momentum:', args.momentum)
         print('DEBUG: args.shuff:', args.shuff)
         print('DEBUG: args.rnd_seed:', args.rnd_seed)
         print('DEBUG: args.backend:', args.backend)
+        print('DEBUG: args.local_rank:', args.local_rank)
         print('DEBUG: args.no_cuda:', args.no_cuda, '\n')
 
     # Encapsulate the model on the GPU assigned to the current process
-    device = torch.device('cuda' if use_cuda else 'cpu', lrank)
     if use_cuda:
         torch.cuda.set_device(lrank)
 
-    # Dataset
+    # Read training dataset
     train_dataset = imagenet_dataset(args.data_dir)
 
     if is_distributed:
         # Distributed sampler restricts data loading to a subset of the dataset
         # exclusive to the current process
-        train_sampler = torch.utils.data.distributed.DistributedSampler(
-            train_dataset, num_replicas=gwsize, rank=grank,
+        train_sampler = DistributedSampler(
+            train_dataset,  # num_replicas=gwsize, rank=grank,
             shuffle=(args.shuff and args.rnd_seed is None)
         )
 
-        train_loader = torch.utils.data.DataLoader(
+        train_loader = DataLoader(
             train_dataset, batch_size=args.batch_size,
             sampler=train_sampler, num_workers=args.nworker, pin_memory=True,
             persistent_workers=(args.nworker > 1),
@@ -172,24 +175,35 @@ def main():
             worker_init_fn=seed_worker
         )
     else:
-        train_loader = torch.utils.data.DataLoader(
+        train_loader = DataLoader(
             train_dataset, batch_size=args.batch_size, generator=torch_prng,
             worker_init_fn=seed_worker
         )
 
     # Create CNN model
-    model = torchvision.models.resnet152().to(device)
+    model = torchvision.models.resnet152()
 
-    # Distribute model to workers
-    if is_distributed:
-        model = nn.parallel.DistributedDataParallel(
-            model,
-            device_ids=[device],
-            output_device=device)
-
-    # Optimizer
-    optimizer = torch.optim.SGD(
-        model.parameters(), lr=args.lr, momentum=args.momentum)
+    # Initialize DeepSpeed to use the following features
+    # 1) Distributed model
+    # 2) DeepSpeed optimizer
+    # 3) Distributed data loader
+    deepspeed_config = {
+        # "train_micro_batch_size_per_gpu": args.batch_size,
+        "optimizer": {
+            "type": "SGD",
+            "params": {
+                "lr": args.lr,
+                "momentum": 0.5
+            }
+        },
+        "fp16": {
+            "enabled": False
+        },
+        "zero_optimization": False
+    }
+    distrib_model, optimizer, deepspeed_train_loader, _ = deepspeed.initialize(
+        args=args, model=model, model_parameters=model.parameters(),
+        training_data=train_dataset, config_params=deepspeed_config)
 
     # Start training loop
     if grank == 0:
@@ -198,8 +212,8 @@ def main():
         print('--------------------------------------------------------')
         nnod = os.environ.get('SLURM_NNODES', 'unk')
         epoch_time_tracker = EpochTimeTracker(
-            series_name="ddp-bl",
-            csv_file=f"epochtime_ddp-bl_{nnod}N.csv"
+            series_name="deepspeed-bl",
+            csv_file=f"epochtime_deepspeed-bl_{nnod}N.csv"
         )
 
     et = timer()
@@ -210,8 +224,11 @@ def main():
             # Inform the sampler that a new epoch started: shuffle
             # may be needed
             train_sampler.set_epoch(epoch)
-        train(model, device, train_loader,
-              optimizer, epoch, grank, gwsize, args)
+
+        # Training
+        train(args, distrib_model, train_loader,
+              optimizer, epoch, grank, gwsize)
+
         # Save first epoch timer
         if epoch == start_epoch:
             first_ep_t = timer()-lt
@@ -224,12 +241,12 @@ def main():
             print('TIMER: epoch time:', timer()-lt, 's')
             epoch_time_tracker.add_epoch_time(epoch-1, timer()-lt)
 
-    if is_distributed:
+    if torch.cuda.is_available():
         dist.barrier()
 
     if grank == 0:
         print('\n--------------------------------------------------------')
-        print('DEBUG: training results:\n')
+        print('DEBUG: results:\n')
         print('TIMER: first epoch time:', first_ep_t, ' s')
         print('TIMER: last epoch time:', timer()-lt, ' s')
         print('TIMER: average epoch time:', (timer()-et)/args.epochs, ' s')
@@ -251,8 +268,7 @@ def main():
 
     # Clean-up
     if is_distributed:
-        dist.barrier()
-        dist.destroy_process_group()
+        deepspeed.sys.exit()
 
 
 if __name__ == "__main__":
