+++ tutorials/distributed-ml/torch-scaling-test/horovod-config.yaml
@@ -1,12 +1,12 @@
 # I/O
-data_dir: tmp_data/
+data_dir: tmp_data/ # /p/scratch/intertwin/datasets/imagenet/ILSVRC2012/train/ # tmp_data/
 log_int: 10
 verbose: True
 nworker: 4 # num workers dataloader
 prefetch: 2
 
 # Model
-batch_size: 64
+batch_size: 64 # micro batch size
 epochs: 3
 lr: 0.001
 momentum: 0.5
